requirements.txt
django
djangorestframework
flake8
mysqlclient
pytz
pyjwt
whitenoise
gunicorn
dj-database-url
psycopg2
Pillow
fuzzywuzzy
